requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
colorama==0.4.6
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
django-phonenumber-field==7.0.2
django-taggit==3.1.0
djangorestframework==3.14.0
drf-nested-routers==0.93.4
frozenlist==1.3.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Markdown==3.4.1
msgpack==1.0.5
multidict==6.0.4
phonenumbers==8.13.7
Pillow==9.4.0
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.1.1
pytz==2022.7.1
redis==4.5.4
requests==2.30.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tqdm==4.65.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==2.0.2
yarl==1.9.2
zope.interface==6.0
